+++ data-structure/dual_segment_tree.cpp
@@ -9,20 +9,20 @@ template <typename M>
 struct DualSegmentTree {
     using T = typename M::T;
 
-    DualSegmentTree(size_t n) {
+    DualSegmentTree(int n) {
         size = 1;
         height = 1;
         while (size < n) size <<= 1, height++;
         lazy.resize(2 * size, M::id);
     }
 
-    T operator[](size_t k) {
+    T operator[](int k) {
         k += size;
         propagate(k);
         return lazy[k];
     }
 
-    void update(size_t l, size_t r, const T& x) {
+    void update(int l, int r, const T& x) {
         l += size;
         r += size;
         propagate(l);
@@ -34,18 +34,18 @@ struct DualSegmentTree {
     }
 
 private:
-    size_t size, height;
+    int size, height;
     vector<T> lazy;
 
-    void push(size_t k) {
+    void push(int k) {
         if (lazy[k] == M::id) return;
         lazy[2 * k] = M::op(lazy[2 * k], lazy[k]);
         lazy[2 * k + 1] = M::op(lazy[2 * k + 1], lazy[k]);
         lazy[k] = M::id;
     }
 
-    void propagate(size_t k) {
-        for (size_t i = height; i > 0; i--) push(k >> i);
+    void propagate(int k) {
+        for (int i = height; i > 0; i--) push(k >> i);
     }
 };
